Add a TRIM function to tink.sql.expr.Functions. Mainly to allow hacking an Int type to a String
A scenario where this gets useful is when we need to map an Int column to a String in a `.next()`.
E.g.

```haxe
// Without the TRIM:
db.Foo
  .select( {
     id    : Foo.id,
     b     : Foo.b,
     c     : Foo.c,
     d     : Foo.d,
     e     : Foo.e,
     f     : Foo.f
   })
   .all()
   .next( o -> {
       id: Std.string(id).map(s -> someTransformation(s)),
       name: name,
       b: b,
       c: c,
       d: d,
       e: e,
       f: f
    });

// Above is very tedious,
// By contrast if we had a TRIM function:
db.Foo
  .select( {
     ref   : Foo.id,
     b     : Foo.b,
     c     : Foo.c,
     d     : Foo.d,
     e     : Foo.e,
     f     : Foo.f
   })
   .all()
   .next( o -> { o.ref = someTransformation(o.ref); o; });

// And no creation of another object, because the Promised object
already types ref as a String.

-------------------------------------------------------

Maybe TRIM is a weird choice here, but I thought having a STRING methods
simply returning an Expr<String> would be more confusing (because it is
not a SQL function).

Tell me if you think there is a better solution :)